[ Feat ] user profile exception

+++ src/main/java/com/backend/simya/domain/profile/service/ProfileService.java
@@ -32,7 +32,7 @@ public class ProfileService {
             Long profileId = profileRepository.save(profile).getProfileId();
             return new ProfileResponseDto(profileId, profile.getNickname(), profile.getUser());
         } catch (Exception ignored) {
-            throw new BaseException(DATABASE_ERROR);
+            throw new BaseException(POST_FAIL_PROFILE);
         }
 
     }
@@ -52,17 +52,27 @@ public class ProfileService {
     // 대표 프로필은 하나만 지정 가능
     public void setMainProfile(Long profileId) throws BaseException {
         Profile profile = getProfileInfo(profileId);
-        User user = profile.getUser();
 
-        int idx = user.getMainProfile();  // 이미 대표 프로필이 존재한다면 해제 후 선택
-        log.info("현재 유저의 main profile index: {}", idx);
+        try {
+            User user = profile.getUser();
+            if (user == null) {
+                throw new BaseException(GET_FAIL_USERINFO);
+            }
 
-        if (idx == -1) {  // 대표 프로필로 지정된 것이 없다면 바로 선택
-            profile.selectMainProfile();
-            log.info("변경 후 main profile index: {}", user.getMainProfile());
-        } else {
-            user.getProfileList().get(idx).cancelMainProfile();
-            profile.selectMainProfile();
+            int idx = user.getMainProfile();  // 이미 대표 프로필이 존재한다면 해제 후 선택
+            log.info("현재 유저의 main profile index: {}", idx);
+
+            if (idx == -1) {  // 대표 프로필로 지정된 것이 없다면 바로 선택
+                profile.selectMainProfile();
+                log.info("변경 후 main profile index: {}", user.getMainProfile());
+            } else {
+                user.getProfileList().get(idx).cancelMainProfile();
+                profile.selectMainProfile();
+            }
+        } catch (IndexOutOfBoundsException exception) {
+            throw new BaseException(USERS_NEED_ONE_MORE_PROFILE);
+        } catch (Exception ignored) {
+            throw new BaseException(SET_FAIL_MAIN_PROFILE);
         }
     }
 
@@ -70,7 +80,14 @@ public class ProfileService {
     public void deleteProfile(Long profileId) throws BaseException {
         try {
             Profile profile = getProfileInfo(profileId);
-            profile.delete(profileId);
+            if (profile.isActivated()) {
+                if (profile.getUser().getProfileList().isEmpty()) {
+                    throw new BaseException(USERS_NEED_ONE_MORE_PROFILE);
+                }
+                profile.delete(profileId);
+            } else {
+                throw new BaseException(ALREADY_DELETE_PROFILE);
+            }
         } catch (Exception ignored) {
             throw new BaseException(DELETE_FAIL_PROFILE);
         }
